tune logback config, replace X-request-id with oculus-request-id
Summary:
Now the log line is too long, try to shrink the length when
oculus-request-id is not available.
With the color, even if log fields are not aligned, it is still
easy to read, so I don't think it would be an issue.

Changes:
- replace all the X-request-id with oculus-request-id
- change httpd log as well, to log oculus-request-id
- increase thread name max length
- make thread/oculus-request-id min length = 0

Our current logstash config can handle the modified format with no error.
Let me know what you guys think.

Test Plan: supercycle+manual

Reviewers: zhany, kg, shuz

Reviewed By: shuz

Subscribers: all

Differential Revision: https://code.silkcloud.info/D1296